refactor!: migrated auth types to the auth feature folder
