+++ server/src/viewmodels/schema.types.ts
@@ -9,9 +9,6 @@ export type ProgrammingLanguageTable = typeof schema.ProgrammingLanguage;
 export type ProjectFrameworksTable = typeof schema.ProjectFramework;
 export type ProjectImagesTable = typeof schema.ProjectImage;
 export type ProjectsTable = typeof schema.Project;
-export type RolesTable = typeof schema.Role;
-export type UsersTable = typeof schema.User;
-export type UserRolesTable = typeof schema.UserRole;
 
 // infer select
 export type ApplicationIndustryViewModel =
@@ -26,9 +23,3 @@ export type ProjectFrameworkViewModel =
   typeof schema.ProjectFramework.$inferSelect;
 export type ProjectImageViewModel = typeof schema.ProjectImage.$inferSelect;
 export type ProjectViewModel = typeof schema.Project.$inferSelect;
-export type RoleViewModel = typeof schema.Role.$inferSelect;
-export type UserViewModel = typeof schema.User.$inferSelect;
-export type UserRoleViewModel = typeof schema.UserRole.$inferSelect;
-
-// infer insert
-export type NewUser = typeof schema.User.$inferInsert;
